// src/types/resources/HealthcareService.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TNarrative } from '../partials/Narrative';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TReference } from '../partials/Reference';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TMarkdown } from '../simpleTypes/Markdown';
import { TAttachment } from '../partials/Attachment';
import { TContactPoint } from '../partials/ContactPoint';
import { THealthcareServiceEligibility } from '../partials/HealthcareServiceEligibility';
import { THealthcareServiceAvailableTime } from '../partials/HealthcareServiceAvailableTime';
import { THealthcareServiceNotAvailable } from '../partials/HealthcareServiceNotAvailable';

export type THealthcareService = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  text?: TNarrative;
  contained?: TResourceContainer[];
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  identifier?: TIdentifier[];
  active?: boolean;
  providedBy?: TReference;
  category?: TCodeableConcept[];
  type?: TCodeableConcept[];
  specialty?: TCodeableConcept[];
  location?: TReference[];
  name?: string;
  comment?: string;
  extraDetails?: TMarkdown;
  photo?: TAttachment;
  telecom?: TContactPoint[];
  coverageArea?: TReference[];
  serviceProvisionCode?: TCodeableConcept[];
  eligibility?: THealthcareServiceEligibility[];
  program?: TCodeableConcept[];
  characteristic?: TCodeableConcept[];
  communication?: TCodeableConcept[];
  referralMethod?: TCodeableConcept[];
  appointmentRequired?: boolean;
  availableTime?: THealthcareServiceAvailableTime[];
  notAvailable?: THealthcareServiceNotAvailable[];
  availabilityExceptions?: string;
  endpoint?: TReference[];
};